requirements.txt
aniso8601==8.0.0
click==7.1.1
ecdsa==0.15
Flask==1.1.1
Flask-API==2.0
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
pyasn1==0.4.8
pymongo==3.10.1
python-dotenv==0.12.0
python-jose==3.1.0
pytz==2019.3
rsa==4.0
six==1.14.0
Werkzeug==1.0.0
WTForms==2.2.1
